requirements.txt
Flask==2.0.1
gunicorn==20.1.0
docutils==0.15
pandas==1.4.4
snowflake-connector-python==2.8.0
tk==0.1.0